% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimX.R
\docType{data}
\name{SimX}
\alias{SimX}
\title{Simulated Data}
\format{
An object of class \code{list} of length 80.
}
\usage{
data(SimX)
}
\description{
A simulated dataset with 2 groups and 80 observations. Each group consists of 40 observations, 5 of which are outliers.
}
\keyword{datasets}
